requirements.txt
streamlit
torch
transformers
chromadb
pypdf2
python-docx
langchain
pyngrok